refactor: :art: created Component RenderButtons to CategoryButtons Component
