+++ src/common/components/CategoryButton.jsx
@@ -1,72 +1,37 @@
-import React, { useContext } from 'react';
+import React from 'react';
+import { FaBookOpen } from 'react-icons/fa';
 import PropTypes from 'prop-types';
-import { FaBookOpen, FaUtensils, FaBreadSlice,
-  FaDrumstickBite, FaIceCream, FaMitten } from 'react-icons/fa';
-
-import store from '../../context/store';
-
-const foodDrinkButtons = [{ strCategory: 'Food' }, { strCategory: 'Drink' }];
+import RenderButtons from './RenderButtons';
 
 export default function CategoryButton({ clickCategory, foodOrDrink, clickAll, path }) { // Desestruturação de props
-  const { recipes: { foods, categoriesMeals,
-    categoriesDrinks, categoriesLimit } } = useContext(store);
-
-  const renderButtons = () => {
-    const mealsBtns = [<FaUtensils key={ 0 } />, <FaBreadSlice key={ 1 } />,
-      <FaDrumstickBite key={ 2 } />, <FaIceCream key={ 3 } />, <FaMitten key={ 4 } />];
-
-    let newCategories;
-    if (path) {
-      newCategories = foodDrinkButtons;
-    } else {
-      newCategories = (foods) ? (
-        categoriesMeals.slice(0, categoriesLimit)) : (
-        categoriesDrinks.slice(0, categoriesLimit));
-    }
-
-    return (
-      newCategories.map((category, index) => (
-        <div key={ index } className="categoriesBtns">
-          <button
-            type="button"
-            data-testid={ path
-              ? `filter-by-${category.strCategory.toLowerCase()}-btn`
-              : `${category.strCategory}-category-filter` }
-            onClick={ path
-              ? (() => foodOrDrink(category.strCategory))
-              : (() => clickCategory(category)) }
-          >
-            {mealsBtns[index]}
-            {category.strCategory}
-          </button>
-        </div>
-      ))
-    );
-  };
-
   return (
     <div className="categoriesBtns">
-      <FaBookOpen />
       <button
         type="button"
         data-testid={ path ? 'filter-by-all-btn' : 'All-category-filter' }
         onClick={ clickAll }
       >
+        {(!path) && <FaBookOpen />}
         All
       </button>
-      {renderButtons()}
+      <RenderButtons
+        clickCategory={ clickCategory }
+        foodOrDrink={ foodOrDrink }
+        path={ path }
+      />
     </div>
   );
 }
 
 CategoryButton.propTypes = {
-  clickCategory: PropTypes.func.isRequired,
+  clickCategory: PropTypes.func,
   clickAll: PropTypes.func.isRequired,
   foodOrDrink: PropTypes.func,
   path: PropTypes.bool,
 };
 
 CategoryButton.defaultProps = {
+  clickCategory: undefined,
   foodOrDrink: () => console.log('nothing to do!'),
   path: false,
 };
